Rename components to prepare views for Consensus transactions
