+++ .changelog/1249.trivial.md
@@ -1,0 +1,1 @@
+Rename components to prepare views for Consensus transactions
