+++ src/app/pages/RuntimeTransactionDetailPage/CurrentFiatValue.tsx
