+++ src/app/pages/RuntimeTransactionDetailPage/__tests__/CurrentFiatValue.test.tsx
